refactor: align schemas with Delta Lake unsigned integer limitation
- Update design.md to document Delta Lake limitation and change all u16/u32 to i16/i32
- Update dim_symbol SCHEMA: exchange_id UInt16 -> Int16
- Remove all dynamic schema reading logic (actual_schema, current_schema)
- Make schema definitions single source of truth
- Update delta_manifest_repo to use explicit Int32 schema (no dynamic casting)
- Update dim_symbol.resolve_symbol_ids to use SCHEMA directly
- Add comprehensive documentation comments about Delta Lake type mapping
- Keep UInt8 for side/asset_type (supported by Delta Lake)

This ensures consistency between code definitions and Delta Lake storage,
eliminating schema mismatch issues that caused manifest merge failures.

+++ tests/test_trades.py
@@ -166,8 +166,8 @@ def test_normalize_trades_schema():
     normalized = normalize_trades_schema(df)
     
     assert normalized["date"].dtype == pl.Date
-    assert normalized["exchange_id"].dtype == pl.UInt16
-    assert normalized["symbol_id"].dtype == pl.UInt32
+    assert normalized["exchange_id"].dtype == pl.Int16  # Delta Lake stores as Int16 (not UInt16)
+    assert normalized["symbol_id"].dtype == pl.Int64  # Delta Lake stores as Int64
     assert normalized["ts_local_us"].dtype == pl.Int64
